backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
boto3==1.34.0
django-storages==1.14.2
requests==2.31.0
openai==1.6.1
gunicorn==21.2.0

# Additional production dependencies
drf-yasg==1.21.7
django-filter==23.5
channels==4.0.0
channels-redis==4.1.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
whitenoise==6.6.0
sentry-sdk==1.39.1
django-debug-toolbar==4.2.0
pyotp==2.9.0
qrcode==7.4.2
django-ratelimit==4.1.0